Bug 707434: Compute offset to bmp pixmap sample as size_t, not int32_t.
When the sizes of bmp images get large enough the computation of the
offset from the beginning of the pixmap samples to the sample that
need to be accessed would overflow, causing the pointer to go beyond
the boundaries of the pixmap samples.

Now we compute the offset as a size_t.

+++ source/fitz/load-bmp.c
@@ -627,7 +627,7 @@ bmp_read_bitmap(fz_context *ctx, struct info *info, const unsigned char *begin, 
 	dstride = pix->stride;
 	if (!info->topdown)
 	{
-		ddp = pix->samples + (height - 1) * dstride;
+		ddp = pix->samples + ((size_t) (height - 1)) * ((size_t) dstride);
 		dstride = -dstride;
 	}
 
@@ -657,8 +657,8 @@ bmp_read_bitmap(fz_context *ctx, struct info *info, const unsigned char *begin, 
 
 	for (y = 0; y < height; y++)
 	{
-		const unsigned char *sp = ssp + y * sstride;
-		unsigned char *dp = ddp + y * dstride;
+		const unsigned char *sp = ssp + ((size_t) y) * ((size_t) sstride);
+		unsigned char *dp = ddp + ((size_t) y) * ((size_t) dstride);
 
 		switch (bitcount)
 		{
